Refactor chat mutation controllers into single-route classes
